Move merge testing, to an internal test file Change merge func to be private Modify comments
