+++ compare/compare.go
@@ -8,11 +8,11 @@ const (
 	Equal Result = iota + 1
 	// BothNulls represents a case in which the current & the other data objects
 	// are null, e.g. null == null
-	// Note: The data is the specific data in the column, and not the whole row
+	// Note: The null refers to both Rows (the Value inside the Data)
 	BothNulls
 	// Null represents a case in which only 1 of the data objects is null,
 	// e.g. 1 == null
-	// Note: The data is the specific data in the column, and not the whole row
+	// Note: The null refers to the Row (the Value inside the Data)
 	Null
 	// Greater represents greater than other data object, e.g. 2 > 1
 	Greater
